Fix for problem with video not finding its "src" attribute.
Did you know that you must use getAttributeNS if ANY attribute has a namespace - not just if the particular attribute you are looking up has a namespace?   Otherwise, it is sometimes not found (particularly if the element has more than a couple of attributes).  This is due to the binary search through attributes that getAttribute uses which does not handle the case of namespaced attributes changing the sorting.